+++ src/components/index.js
@@ -6,4 +6,5 @@ iconLibrary.add(fas, fab)
 
 export { Button } from './buttons/Button/Button.jsx'
 export { ThemeProvider } from './contexts/ThemeContext.js'
+export { HeaderLink } from './buttons/HeaderLink.jsx'
 export { Header } from './layout/Header/Header.jsx'